cleanup explorer hiding logic
